% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\title{BLA}
\usage{
decode(encoded)
}
\arguments{
\item{encoded}{character, encoded flexible polyline string.}
}
\value{
String
}
\description{
BLA
}
\examples{
# 2D
decode("BFoz5xJ67i1B1B7PzIhaxL7Y")

# 3D
decode("BlBoz5xJ67i1BU1B7PUzIhaUxL7YU")
# routes$geometry[1] \%>\% decode() \%>\% st_linestring() \%>\% st_sfc(crs = 4326) \%>\% mapview()
}
