% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode_sf}
\alias{decode_sf}
\title{Wrapper function for decoding to simple features}
\usage{
decode_sf(encoded, crs = 4326)
}
\arguments{
\item{encoded}{crs, coordinate reference system (\code{default = 4326}, WGS84).}
}
\value{
A matrix containing the coordinates of the decoded line.
}
\description{
A wrapper function for \code{\link{decode}} that converts the input polylines
to simple feature geometries of the sf package.
This function calls \code{hf::polyline_decode} and \code{hf::get_third_dimension}
of the C++ implementation of the flexible polyline encoding by HERE. Depending
on the dimensions of the encoded line, a two or three dimensional line is decoded.
}
\note{
The function returns a sf object, therefore the input set of polylines must
be of consistent dimension (e.g XY or XYZ) to meet the requirements of the
constructor of an sf object. For mixed dimensions use the \code{\link{decode}}
function directly.
}
\examples{
# 2d line
decode_sf("BFoz5xJ67i1B1B7PzIhaxL7Y")

# 3d line
decode_sf("BlBoz5xJ67i1BU1B7PUzIhaUxL7YU")
}
