% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexpolyline-package.R
\docType{package}
\name{flexpolyline-package}
\alias{flexpolyline}
\alias{flexpolyline-package}
\title{flexpolyline: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
