% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\title{Decode a flexible polyline encoded string}
\usage{
decode(encoded)
}
\arguments{
\item{encoded}{character, encoded flexible polyline string.}
}
\value{
A matrix containing the coordinates.
}
\description{
Decode a flexible polyline encoded string
}
\examples{
# 2d line
decode("BFoz5xJ67i1B1B7PzIhaxL7Y")

# 3d line
decode("BlBoz5xJ67i1BU1B7PUzIhaUxL7YU")
# routes$geometry[1] \%>\% decode() \%>\% st_linestring() \%>\% st_sfc(crs = 4326) \%>\% mapview()
}
