% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode}
\alias{encode}
\title{Encode a line in flexible polyline encoding}
\usage{
encode(line, precision = 5)
}
\arguments{
\item{line}{matrix, coordinates of the line in 2d or 3d.}

\item{precision}{integer, precision to use in encoding (between 1 and 15; only valid results until 7!).}
}
\value{
The line as encoded string.
}
\description{
Encode a line in flexible polyline encoding
}
\examples{
# 2D
line2d <- matrix(
  c(8.69821, 50.10228,
    8.69567, 50.10201,
    8.69150, 50.10063,
    8.68752, 50.09878),
  ncol = 2, byrow = TRUE
)
encode(line2d)

# 3D
line3d <- matrix(
  c(8.69821, 50.10228, 10,
    8.69567, 50.10201, 20,
    8.69150, 50.10063, 30,
    8.68752, 50.09878, 40),
  ncol = 3, byrow = TRUE
)
encode(line3d)
}
