% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexpolyline-package.R
\docType{package}
\name{flexpolyline-package}
\alias{flexpolyline}
\alias{flexpolyline-package}
\title{flexpolyline: Flexible Polyline Encoding}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Binding to the C++ implementation of the flexible polyline
    encoding by HERE <https://github.com/heremaps/flexible-polyline>. The
    flexible polyline encoding is a lossy compressed representation of a list of
    coordinate pairs or coordinate triples. The encoding is achieved by:
    (1) Reducing the decimal digits of each value;
    (2) encoding only the offset from the previous point;
    (3) using variable length for each coordinate delta; and
    (4) using 64 URL-safe characters to display the result.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://munterfinger.github.io/flexpolyline/}
  \item \url{https://github.com/munterfinger/flexpolyline/}
  \item Report bugs at \url{https://github.com/munterfinger/flexpolyline/issues/}
}

}
\author{
\strong{Maintainer}: Merlin Unterfinger \email{info@munterfinger.ch} (\href{https://orcid.org/0000-0003-2020-2366}{ORCID})

Authors:
\itemize{
  \item  HERE Europe B.V. (Flexible polyline encoding C++ implementation) [copyright holder]
}

}
\keyword{internal}
