% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexpolyline-package.R
\docType{package}
\name{flexpolyline-package}
\alias{flexpolyline}
\alias{flexpolyline-package}
\title{flexpolyline: Flexible polyline encoding and decoding}
\description{
The flexible polyline encoding is a lossy compressed representation
    of a list of coordinate pairs or coordinate triples. Provides a binding to
    the C++ implementation of the encoding by HERE <https://github.com/heremaps/flexible-polyline>.
}
\author{
\strong{Maintainer}: Merlin Unterfinger \email{info@munterfinger.ch}

}
\keyword{internal}
